# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Country pass API documentation",
        "description": "Walk between countries, if its available",
        "contact": {
            "email": "me@kalanys.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:60000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/routing/{country1}/{country2}": {
            "get": {
                "tags": [
                    "Targeting"
                ],
                "summary": "Get available addresses by sent parts",
                "operationId": "dce4aba6d87c697631e8a7b50f7251e5",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Available path",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "No path found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "419": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Targeting",
            "description": "Targeting paths via known data"
        }
    ]
}